services/api/requirements.txt
fastapi==0.114.0
uvicorn[standard]==0.30.0
pandas==2.2.2
numpy==1.26.4
statsmodels==0.14.2
prophet==1.1.5
xgboost==2.1.1
tensorflow==2.16.2
google-cloud-bigquery==3.25.0
pydantic==2.8.2
